+++ glotaran/plugin_system/io_plugin_utils.py
@@ -56,10 +56,10 @@ def inferr_file_format(
         return file_format.lstrip(".")
 
     if allow_folder:
-        return "folder"
+        return "legacy"
     else:
         raise ValueError(
-            f"Cannot determine format of file {file_path!r}, " "please provide an explicit format."
+            f"Cannot determine format of file {file_path!r}, please provide an explicit format."
         )