+++ changelog.md
@@ -10,6 +10,7 @@
 ### 👌 Minor Improvements:
 
 - 👌🎨 Add proper repr for DatasetMapping (#957)
+- 👌 Add SavingOptions to save_result API (#966)
 
 ### 🩹 Bug fixes
 
@@ -18,6 +19,8 @@
 ### 📚 Documentation
 
 ### 🗑️ Deprecations (due in 0.8.0)
+
+- `glotaran.io.save_result(result, result_path, format_name='legacy')` -> `glotaran.io.save_result(result, Path(result_path) / 'result.yml')`
 
 ### 🚧 Maintenance